Finish server

+++ EditCarDB__NP/Models/Car.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Make { get; set; }
         public string Model { get; set; }
-        public ushort Year { get; set; }
+        public int Year { get; set; }
         public string Color { get; set; }
     }
 }
